% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{age_prep}
\alias{age_prep}
\title{Pre-calculation of equilibrium age-related values}
\usage{
age_prep(age, eta, rho, a0)
}
\arguments{
\item{age}{Vector of age groups}

\item{eta}{Parameter eta}

\item{rho}{Parameter rho}

\item{a0}{Parameter a0}
}
\value{
A named list of equilibrium age-related values
}
\description{
Pre-calculation of equilibrium age-related values
}
