% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{break_coverage}
\alias{break_coverage}
\title{Break coverage}
\usage{
break_coverage(breaks, range_min, range_max)
}
\arguments{
\item{breaks}{vector of breaks}

\item{range_min}{minimum value of range of interest}

\item{range_max}{maximum value of range of interest}
}
\description{
Feed in a vector of breaks x, and min and max values of a range
  that falls within x. Output the proportion of each slice that is covered by
  the range. Useful for calculating quantities like prevalence in a given age
  range.
}
