% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{breakCoverage}
\alias{breakCoverage}
\title{Break coverage}
\usage{
breakCoverage(breaks, range_min, range_max)
}
\arguments{
\item{breaks}{vector of breaks}

\item{range_min}{minimum value of range of interest}

\item{range_max}{maximum value of range of interest}
}
\description{
Feed in a vector of breaks x, and min and max values of a range that falls within x. Output the proportion of each slice that is covered by the range. Useful for calculating things like prevalence in a given age range.
}
