% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{create_prior_string}
\alias{create_prior_string}
\title{Create a C++ prior function string}
\usage{
create_prior_string(project)
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}).}
}
\description{
Given a project with model and fitting priors already defined,
  creates a C++ function in the form of a string that will return the joint
  density of these priors in log space. This string can then be passed to
  \code{drjacoby} and used in MCMC.
}
