% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Default parameters}
\usage{
default_parameters()
}
\description{
Returns list of default parameters, taken from Griffin et al. 2014 Nature Communications 5, "Estimates of the changing age-burden of Plasmodium falciparum malaria disease in sub-Saharan Africa". All time units are in days.
}
