% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{define_prior}
\alias{define_prior}
\title{Define a prior distribution over a parameter within a project}
\usage{
define_prior(project, name, prior_dist, prior_params, range_min = NULL,
  range_max = NULL)
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}).}

\item{name}{the name of a parameter within the project model parameters or
fitting parameters.}

\item{prior_dist}{the type of prior distribution to define. One of {"fixed",
"dbeta", "dnorm", "dlnorm", "dgamma"}.}

\item{prior_params}{parameters associated with the prior distribution.}

\item{range_min}{the minimum allowed parameter value.}

\item{range_max}{the maximum allowed parameter value.}
}
\description{
Given a model fitting project with model parameters and/or
  fitting parameters loaded, redefine the prior distribution of a named
  parameter. This function will try to find the given parameter name within
  both \code{project$model_params} and \code{project$fitting_params}, hence
  this function can be used to specify priors over both objects.
}
