% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{draw_fitting_prior}
\alias{draw_fitting_prior}
\title{Draw a set of fitting parameters from the prior}
\usage{
draw_fitting_prior(project)
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}), with fitting parameters already loaded.}
}
\description{
Takes a fitting project with fitting priors already loaded, and
  draws a new parameter set from the prior.
}
