% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{draw_model_prior}
\alias{draw_model_prior}
\title{Draw a set of model parameters from the prior}
\usage{
draw_model_prior(project)
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}), with model parameters already loaded.}
}
\description{
Takes a fitting project with model priors already loaded, and
  draws a new parameter set from the prior.
}
