% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_likelihood.R
\name{function_to_string}
\alias{function_to_string}
\title{Extract function as text from .cpp file}
\usage{
function_to_string(address)
}
\arguments{
\item{address}{File address}
}
\value{
function as string
}
\description{
Conversion of .cpp loglikelihood file to string for use with DrJacoby. The 
file provided should contain a single function which returns SEXP and has a name
starting "loglikelihood". The end of the function must be flagged with 
"// loglikelihood_end" on the final line"
}
\examples{
\dontrun{
  An example likelihood file (example.cpp) may look like:
  # include <Rcpp.h>
  # [[Rcpp::export]]
  # SEXP loglikelihood(std::vector<double> params, std::vector<double> x){        
  #   // extract parameters
  #   double mu = params[0];
  #   double sigma = params[1];#'
  #   // sum log-likelihood over all data
  #   double ret = 0.0;
  #   for (size_t i = 0; i < sizeof(x); ++i) {
  #     ret += -0.5*log(2*M_PI*sigma*sigma) - (x[i] - mu)*(x[i] - mu)/(2*sigma*sigma);
  #    }
  #   // return as SEXP
  #   return Rcpp::wrap(ret);
  # }
  # // loglikelihood_end

  Add we could convert it by:
  # function_to_string("src/example.cpp")
}
}
