% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_inc}
\alias{get_inc}
\title{Get incidence from equilibrium output}
\usage{
get_inc(eq, age0, age1)
}
\arguments{
\item{eq}{equilibrium solution, as returned by \code{human_equilibrium()}.}

\item{age0}{lower bound on age range.}

\item{age1}{upper bound on age range.}
}
\description{
Get incidence in a given age range, using pre-computed
  equilibrium solution as input.
}
