% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_prev}
\alias{get_prev}
\title{Get prevalence from equilibrium output}
\usage{
get_prev(eq, age0, age1, prev_type = "microscopy")
}
\arguments{
\item{eq}{equilibrium solution, as returned by \code{human_equilibrium()}.}

\item{age0}{lower bound on age range.}

\item{age1}{upper bound on age range.}

\item{prev_type}{whether to calculate prevalence by "microscopy" or "PCR".}
}
\description{
Get prevalence by microscopy or PCR in a given age range, using
  pre-computed equilibrium solution as input.
}
