% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{human_equilibrium_no_het}
\alias{human_equilibrium_no_het}
\title{Equilibrium solution without biting heterogeneity}
\usage{
human_equilibrium_no_het(EIR, ft, p, age)
}
\arguments{
\item{EIR}{EIR for adults}

\item{ft}{proportion of clinical cases effectively treated}

\item{p}{vector of model parameters}

\item{age}{vector of age groups}
}
\description{
Returns the equilibrium states for the model of Griffin et al.
  2014 Nature Communications 5. "Estimates of the changing age-burden of
  Plasmodium falciparum malaria disease in sub-Saharan Africa". For
  derivation of equilibrium solutions see Griffin 2016 Malaria Journal "Is a
  reproduction number of one a threshold for Plasmodium falciparum malaria
  elimination?" supplementary material.
  
  This function does not account for biting heterogeneity - see 
  human_equilibrium() for function that takes this into account.
  
  Original code due to Jamie Griffin, later modified by Xiaoyu Liu and Bob 
  Verity.
}
