% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{human_equilibrium_no_het_compiled}
\alias{human_equilibrium_no_het_compiled}
\title{Compiled human_equilibrium_no_het}
\usage{
human_equilibrium_no_het_compiled(EIR, ft, p, age)
}
\arguments{
\item{EIR}{EIR for adults}

\item{ft}{proportion of clinical cases effectively treated}

\item{p}{vector of model parameters}

\item{age}{vector of age groups}
}
\description{
Compiled version of human_equilibrium_no_het() function - see
  original function for details.
}
