% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Griffin.R
\name{human_equilibrium_noHet}
\alias{human_equilibrium_noHet}
\title{Equilibrium solution without biting heterogeneity}
\usage{
human_equilibrium_noHet(EIR, ft, p, age)
}
\arguments{
\item{EIR}{EIR for adults}

\item{ft}{proportion of clinical cases effectively treated}

\item{p}{vector of model parameters}

\item{age}{vector of age groups}

\item{h}{Gauss-Hermite nodes and weights for heterogeneity in biting}
}
\description{
Returns the equilibrium states and R0 for the model of Griffin et al. 2014 Nature Communications 5. "Estimates of the changing age-burden of Plasmodium falciparum malaria disease in sub-Saharan Africa". Does not account for biting heterogeneity - see human_equilibrium() for function that takes this into account.
}
\details{
Original code due to Jamie Griffin, later modified by Xiaoyu Liu and Bob Verity.
}
