% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Griffin.R
\name{human_equilibrium_noHet_compiled}
\alias{human_equilibrium_noHet_compiled}
\title{Compiled human_equilibrium_no_het}
\usage{
human_equilibrium_noHet_compiled(EIR, ft, p, age)
}
\description{
Compiled version of human_equilibrium_no_het() function - see original function for details.
}
