% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{likelihood_string}
\alias{likelihood_string}
\title{Extracted string version of the c++ loglikelihood functon for use in drjacoby}
\format{An object of class \code{character} of length 1.}
\usage{
likelihood_string
}
\description{
Extracted string version of the c++ loglikelihood functon for use in drjacoby
}
\keyword{datasets}
