% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{load_data}
\alias{load_data}
\title{Load data into a project}
\usage{
load_data(project, data_df)
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}).}

\item{data_df}{a dataframe of fitting data, formatted as described in the
details section.}
}
\description{
Perform basic checks on data format (see details), and if
  passed, load into an existing project.
}
\details{
TODO.
}
