% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{load_fitting_priors}
\alias{load_fitting_priors}
\title{Load fitting priors into a project}
\usage{
load_fitting_priors(project, file_name = "refit2020_fitting_priors.rds")
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}).}

\item{file_name}{the name of a file within the inst/extdata/fitting_priors
folder.}
}
\description{
Priors involved in fitting to the data are stored within the
  package inst/extdata/fitting_priors folder. Load one of these objects by
  name, and attach to an existing project.
}
