% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{load_model_priors}
\alias{load_model_priors}
\title{Load model priors into a project}
\usage{
load_model_priors(project, file_name = "refit2020_model_priors.rds")
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}).}

\item{file_name}{the name of a file within the inst/extdata/model_priors
folder.}
}
\description{
Model priors are stored within the package
  inst/extdata/model_priors folder. Load one of these objects by name, and
  attach to an existing project.
}
