% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Load parameter set}
\usage{
load_parameters(name = "parameters_Griffin2014.txt")
}
\arguments{
\item{name}{name of parameter set to load}
}
\description{
Load a named parameter set from the inst/parameter_sets folder
  of this package. Input is filtered to relevant parameters only, and
  returned as a named list.
}
