% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic}
\alias{logistic}
\title{Logistic, i.e. inverse logit transform}
\usage{
logistic(x)
}
\arguments{
\item{x}{single value or vector of values to transform.}
}
\description{
Return logistic transform of x, also known as the inverse-logit
  transform, defined as 1/(1 + exp(-x)).
}
