% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nice1}
\alias{plot_nice1}
\title{plot_nice1}
\usage{
plot_nice1(x, y, col = 1, xlab = NULL, ylab = NULL, main = NULL,
  pch = 1, cex = 1, legend = FALSE, legend_position = "right",
  legend_title = "legend_title", legend_label = NULL)
}
\description{
Simple plotting function for producing plot similar to base plot(), but in ggplot2. Has the advantage of being able to specify legends outside of plotting area more easily.
}
