% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(data, p, output_type = "ll", nh = 5)
}
\arguments{
\item{data}{Standardised data input}

\item{p}{Parameter list}

\item{output_type}{Ouput type: "ll" = Log likelihood, "eq" = Equilibrium}

\item{nh}{Number of heterogeneity groups}
}
\value{
Vector of data in correct format
}
\description{
Prepare (standardised) data inputs for use with Equilibrium/likelihood master function
}
