% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{random_parameters}
\alias{random_parameters}
\title{Draw random parameters}
\usage{
random_parameters(n = 1)
}
\arguments{
\item{n}{nuber of random parameter sets to draw.}
}
\description{
Set some parameters as default, and draw others from
  distributions. TODO - find out where these distributions and values came
  from.
}
