% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{required_eq_params}
\alias{required_eq_params}
\title{Return names of required equilibrium parameters}
\usage{
required_eq_params()
}
\description{
Return names of required equilibrium parameters. Note that this 
  is a subset of all parameters, as some parameters are only needed in a 
  dynamic setting.
}
