% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rinvgamma}
\alias{rinvgamma}
\title{Draw from inverse gamma distribution}
\usage{
rinvgamma(n, shape, scale)
}
\arguments{
\item{n}{number of observations.}

\item{shape, scale}{shape and scale parameters of the distribution.}
}
\description{
Draw from inverse gamma distribution with given shape and scale
  parameters. This is equivalent to 1/X, where X is drawn from
  \code{rgamma(1, shape = shape, rate = scale)} (note that the *rate* of the
  gamma draw becomes the *scale* of the inverse gamma draw).
}
