% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fit.R
\name{run_fit}
\alias{run_fit}
\title{Run model fitting}
\usage{
run_fit(project, ...)
}
\arguments{
\item{project}{A mmfit project containing description, model priors, fitting priors and data}

\item{...}{Additional named parameters passed to the \code{\link[drjacoby]{run_mcmc}} function}
}
\value{
Project with fitting output
}
\description{
Run model fitting
}
