% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_parameters}
\alias{set_parameters}
\title{Set remaining parameters}
\usage{
set_parameters(v)
}
\arguments{
\item{v}{vector of parameters, with names corresponding to those returned by
\code{default_parameters()}.}
}
\description{
Replace some default parameters with values from a parameter
  vector v.
}
