% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_site}
\alias{sim_site}
\title{Simulate data at a single site}
\usage{
sim_site(
  p,
  age0 = c(0, 1, 10),
  age1 = c(1, 5, 20),
  n = c(10, 10, 10),
  clin_p_years = c(10, 10, 10),
  EIR = 5,
  case_detection_method = 1,
  treatment = 0.2,
  sigma_p = 0,
  theta = 0.1,
  sigma_c = 0,
  alpha_c = 1,
  slide_on = TRUE,
  PCR_on = TRUE,
  clin_on = TRUE
)
}
\arguments{
\item{p}{a parameters object of class "mmfit_params" (see
\code{?load_parameter_set()}).}

\item{age0, age1}{lower and upper bounds on observed age ranges.}

\item{n}{when simulating prevalence, the number of individuals tested in each
age group.}

\item{clin_p_years}{when simulating incidence, the time at risk of
individuals in each age group.}

\item{EIR}{the entomological inoculation rate of the site, in units of
infectious bites per person per year.}

\item{case_detection_method}{code for case detection method: \itemize{ \item
1 Daily active case detection (ACD) \item 2 Weekly ACD \item 3 Passive case
detection (PCD) }}

\item{treatment}{treatment rate in population.}

\item{sigma_p}{standard deviation of the normally-distributed study-level
random effect on prevalence.}

\item{theta}{overdispersion parameter of the beta-binomial distribution on
prevalence. 0 = standard binomial, and theta must be <1.}

\item{sigma_c}{standard deviation of the normally-distributed study-level
random effect on incidence.}

\item{alpha_c}{standard deviation of the gamma-distributed site- and
age-level random effect on incidence.}

\item{slide_on}{whether to report results of slide prevalence.}

\item{PCR_on}{whether to report results of PCR prevalence.}

\item{clin_on}{whether to report results of clinical incidence.}
}
\description{
Simulate data in the same format as real world data by drawing
  from the likelihood framework. This function simulates data at a single
  study and single site (arbitrarily given a value of 1), see
  \code{?sim_study()} and \code{?sim_complete()} for alternatives.
}
