% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_study}
\alias{sim_study}
\title{TODO - title}
\usage{
sim_study(sigma_c = 1, sigma_p = 1)
}
\arguments{
\item{sigma_c}{standard deviation of the normally-distributed study-level
random effect on incidence.}

\item{sigma_p}{standard deviation of the normally-distributed study-level
random effect on prevalence.}
}
\description{
TODO - description
}
\details{
TODO - details
}
