% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Griffin.R
\name{find_inc}
\alias{find_inc}
\title{find_inc}
\usage{
find_inc(eq, age0, age1)
}
\arguments{
\item{eq}{equilibrium solution, as returned by human_equilibrium()}

\item{age0}{lower bound on age range over which to calculate prevalence}

\item{age1}{upper bound on age range over which to calculate prevalence}
}
\description{
Find incidence in a given age group. Uses pre-computed equilibrium solution as input.
}
\details{
Original code due to Jamie Griffin, later modified by Xiaoyu Liu and Bob Verity.
}
