% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Griffin.R
\name{find_prev}
\alias{find_prev}
\title{find_prev}
\usage{
find_prev(eq, age0, age1, prevType)
}
\arguments{
\item{eq}{equilibrium solution, as returned by human_equilibrium()}

\item{age0}{lower bound on age range over which to calculate prevalence}

\item{age1}{upper bound on age range over which to calculate prevalence}

\item{prevType}{whether to calculate prevalence by "microscopy" or "PCR"}
}
\description{
Find prevalence by microscopy or PCR in a given age group. Reads in equilibrium solution as input.
}
\details{
Original code due to Jamie Griffin, later modified by Xiaoyu Liu and Bob Verity.
}
