% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{load_data}
\alias{load_data}
\title{Load data into a project}
\usage{
load_data(project, file_name = "refit2020_data.rds")
}
\arguments{
\item{project}{an object of class "mmfit_project" (see
\code{?mmfit_project()}).}

\item{file_name}{the name of a file within the inst/extdata/data folder.}
}
\description{
Data for use in model fitting is stored within the package
  inst/extdata/data folder. Load a data object from this folder by name, and
  attach to an existing project.
}
