% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaModelFit.R
\docType{package}
\name{malariaModelFit}
\alias{malariaModelFit}
\alias{malariaModelFit-package}
\title{Fits the Griffin et. al. (2014) malaria transmission model}
\description{
Contains functions for re-fitting the Griffin et al. (2014)
  model to data using MCMC.
}
\references{
Griffin et. al. (2014). Estimates of the changing age-burden of
  Plasmodium falciparum malaria disease in sub-Saharan Africa.
  doi:10.1038/ncomms4136
}
