% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{malariaModelFit_file}
\alias{malariaModelFit_file}
\title{Load sytem file for this package}
\usage{
malariaModelFit_file(name)
}
\arguments{
\item{name}{name of file}
}
\description{
Load and return file from within the inst folder of this
  package.
}
