% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(data, eta = 0.0001304631, rho = 0.85, a0 = 2920,
  output_type = "ll", nh = 5)
}
\arguments{
\item{data}{Standardised data input}

\item{output_type}{Ouput type: "ll" = Log likelihood, "eq" = Equilibrium}

\item{nh}{Number of heterogeneity groups}

\item{p}{Parameter list}
}
\value{
Vector of data in correct format
}
\description{
Prepare (standardised) data inputs for use with Equilibrium/likelihood master function
}
