% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_site}
\alias{sim_site}
\title{Simulate data at a single site}
\usage{
sim_site(
  p,
  EIR = 5,
  treatment = 0.2,
  type = 1,
  age0 = 0:4,
  age1 = 1:5,
  denom = rep(10, 5),
  case_detection = 1,
  theta = 0.1,
  alpha_c = 1,
  u = 0,
  w = 1
)
}
\arguments{
\item{p}{a parameters object of class "model_params" (see
\code{?load_parameter_set()}).}

\item{EIR}{the entomological inoculation rate of the site, in units of
infectious bites per person per year.}

\item{treatment}{treatment rate in population.}

\item{type}{the data type to be simulated:
\itemize{
  \item 1 clinical incidence, where \code{denom} is the person-years at risk
  and \code{numer} becomes the number of cases in that period.
  \item 2 prevalence by microscopy, where \code{denom} is the number of
  people tested and \code{numer} becomes the number of positive tests.
}}

\item{age0, age1}{lower and upper bounds on observed age ranges.}

\item{denom}{the denominator used in simulation. When simulating prevalence
this represents the number of individuals tested in each age group, when
simulating incidence this represents the time at risk of individuals in
each age group.}

\item{case_detection}{code for case detection method:
\itemize{
  \item 1 Daily active case detection (ACD), detection probability 1.0.
  \item 2 Weekly ACD, detection probability 0.5.
  \item 3 Passive case detection (PCD), detection probability 0.2.
}}

\item{theta}{overdispersion parameter of the beta-binomial distribution on
prevalence. 0 = standard binomial, and theta must be <1.}

\item{alpha_c}{variance of the gamma-distributed age-level random effect on
incidence for this site.}

\item{u, w}{study-level random effect on incidence and prevalence,
respectively. Leave at 0 when simulating a single site.}
}
\description{
Simulate data in the same format as real world data by drawing
  from the likelihood framework. This function simulates data at a single
  study and single site (arbitrarily given a value of 1), see
  \code{?sim_study()} and \code{?sim_complete()} for alternatives.
}
